Extract content-store handler into separate file
